chore: move format error to error instance
