+++ packages/core/src/utils/index.ts
@@ -2,4 +2,3 @@ export * from "./parse-schema";
 export * from "./presence";
 export * from "./dot-prop";
 export * from "./object";
-export * from "./format-errors";
